+++ android/src/main/kotlin/io/emma/emma_flutter_sdk/Utils.kt
@@ -1,7 +1,6 @@
 package io.emma.emma_flutter_sdk
 
 import android.content.Context
-import android.content.pm.ApplicationInfo
 import android.content.pm.PackageManager
 import android.content.res.Resources
 import android.os.Handler